fix workout user table & goal

+++ database/migrations/2023_01_28_163638_create_workout_users_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->integer('age')->nullable();
             $table->integer('height')->nullable();
             $table->integer('weight')->nullable();
-            $table->integer('bmi')->nullable();
+            $table->double('bmi')->nullable();
             $table->integer('level')->default(1);
             $table->timestamps();
         });
